Front and Back: add method Jasper PDF and Format in angular styles

+++ corralon/src/main/java/com/undec/corralon/repository/BancoRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface BancoRepository extends JpaRepository <Banco, Integer> {
-    List<Banco>findAllByHabilitadoEquals(Integer habilitado);
+    List<Banco>findAllByHabilitadoEquals(Byte habilitado);
 }
